discojs/src/models: move and export hellaswag
