+++ discojs/src/models/index.ts
@@ -3,5 +3,7 @@ export { BatchLogs, EpochLogs, ValidationMetrics } from "./logs.js";
 
 export { GPT } from './gpt/index.js'
 export { GPTConfig } from './gpt/config.js'
+export { ONNXModel } from './onnx.js'
+export { evaluate as evaluate_hellaswag } from './hellaswag.js'
 export { TFJS } from './tfjs.js'
 export { getTaskTokenizer } from './tokenizer.js'
